fix codigo funciona app agregar con error de catalogo

+++ app/src/main/java/com/leny/yanel/app/api/model/ProductosResponse.kt
@@ -22,7 +22,7 @@ data class Pageable(
     val sort: Sort,
     val pageNumber: Int,
     val pageSize: Int,
-    val offset: Int,
+    val offset: Long,
     val paged: Boolean,
     val unpaged: Boolean
 )
